exception handling - refactored book controller

+++ librarymangement-backend/src/main/java/com/med/library/restController/BookRestController.java
@@ -1,8 +1,12 @@
 package com.med.library.restController;
 
 import com.med.library.entity.Book;
+import com.med.library.restExceptionHandler.exception.HttpNotFoundException;
 import com.med.library.service.BookService;
+
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.http.HttpStatus;
+import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
 
 import java.util.List;
@@ -20,25 +24,38 @@ public class BookRestController {
     }
 
     @GetMapping("/{bookId}")
-    public Book getBookById(@PathVariable("bookId") long bookId) {
-        return bookService.findById(bookId).orElse(null);
+    public ResponseEntity<Book> getBookById(@PathVariable("bookId") long bookId) {
+        Book persistedBook = bookService.findById(bookId);
+        if ( persistedBook == null ) {
+            throw new HttpNotFoundException("ID " + bookId + " not found.");
+        }
+        return  new ResponseEntity<>(persistedBook, HttpStatus.OK);
     }
 
     @PostMapping
     public Book saveBook(@RequestBody Book book) {
-        book.setId(0L);
+        book.setId(0);
         return bookService.save(book);
     }
 
     @PutMapping("/{bookId}")
-    public Book updateBook(@PathVariable("bookId") long bookId  , @RequestBody Book book) {
+    public ResponseEntity<Book> updateBook(@PathVariable("bookId") long bookId  , @RequestBody Book book) {
+        Book persistedBook = bookService.findById(bookId);
+        if ( persistedBook == null ) {
+            throw new HttpNotFoundException("ID " + bookId + " not found.");
+        }
         book.setId(bookId);
-        return bookService.save(book);
+        Book updatedBook = bookService.save(book);
+        return new ResponseEntity<>(updatedBook, HttpStatus.OK);
     }
 
     @DeleteMapping("/{bookId}")
-    public String deleteBook(@PathVariable("{bookId}") long bookId) {
+    public ResponseEntity<?> deleteBook(@PathVariable("bookId") long bookId) {
+        Book persistedBook = bookService.findById(bookId);
+        if ( persistedBook == null ) {
+            throw new HttpNotFoundException("ID " + bookId + " not found.");
+        }
         bookService.deleteById(bookId);
-        return "Book deleted !";
+        return new ResponseEntity<>("Book deleted successfully !", HttpStatus.OK);
     }
 }
